requirements.txt
python-binance
ta
pandas
mplfinance
requests

